refactor: split settings parsing out of types.ts
Move parseSettings and its private helpers (resolveStrippedFields,
filterRequiredFields, resolvePrLabels, isStringArray, isPlainObject)
into a new settings-parse.ts. types.ts is now type declarations plus
DEFAULT_SETTINGS and errorMessage.

Test file renamed to match module ownership.
